Move non-primitive actions into runner
