+++ src/electron/window.js
@@ -3,9 +3,6 @@ let renderer      = electron.ipcMain;
 let urlFormat     = require('url');
 let FrameManager  = require('./frame-manager');
 
-// URL protocols that don't need to be checked for validity
-const KNOWN_PROTOCOLS = ['http', 'https', 'file', 'about', 'javascript'];
-
 
 function Window(win, options){
   this.initialize(win, options);
@@ -70,10 +67,6 @@ Window.prototype.abortPending = function(){
  * navigate {url, headers}
  */
  Window.prototype.navigate = function(url, headers){
-   if (!url || typeof url !== 'string') {
-     let error = new Error('navigate: `url` must be a non-empty string');
-     return Promise.reject(error);
-   }
    if (this.webContents.getURL() == url) {
      return Promise.resolve();
    }
@@ -106,9 +99,6 @@ Window.prototype.abortPending = function(){
          };
        }
      };
-     let handleDomReady = () => {
-       this.domLoaded = true;
-     };
      let handleFinish = () => {
        // We will have already unsubscribed if load failed, so assume success.
        cleanup();
@@ -116,18 +106,15 @@ Window.prototype.abortPending = function(){
        setImmediate(() => resolve(responseData));
      };
      let setup = () => {
-       this.domLoaded = false;
        this.webContents.on('did-fail-load',             handleFailure);
        this.webContents.on('did-fail-provisional-load', handleFailure);
        this.webContents.on('did-get-response-details',  handleDetails);
-       this.webContents.on('dom-ready',                 handleDomReady);
        this.webContents.on('did-finish-load',           handleFinish);
      };
      let cleanup = () => {
        this.webContents.removeListener('did-fail-load',             handleFailure);
        this.webContents.removeListener('did-fail-provisional-load', handleFailure);
        this.webContents.removeListener('did-get-response-details',  handleDetails);
-       this.webContents.removeListener('dom-ready',                 handleDomReady);
        this.webContents.removeListener('did-finish-load',           handleFinish);
        this.setIsReady(true);
      };
@@ -157,54 +144,13 @@ Window.prototype.abortPending = function(){
    });
  };
 
-
-/**
- * goto {url, headers, timeout}
- */
-Window.prototype.goto = function(url, headers, timeout){
-  let timer = wait(timeout)
-    .then(() => {
-      // Even if "successful," note that some things didn't finish.
-      responseData.details = `Not all resources loaded after ${timeout} ms`;
-      //Navigation error
-      let error = {
-        message: 'navigation error',
-        code: -7, // chromium's generic networking timeout code
-        details: `Navigation timed out after ${timeout} ms`,
-        url: url
-      };
-      this.setIsReady(true);
-      // If the DOM loaded before timing out, consider the load successful.
-      return this.domLoaded ? Promise.resolve(responseData) : Promise.reject(error);
-    });
-
-  let goto = canLoadProtocol(url)
-    .then(() => this.abortPending())
-    .then(() => this.navigate(url, headers));
-
-  return Promise.race([timer, goto]);
-};
-
 /**
  * javascript {src}
  */
 Window.prototype.javascript = function(src){
   let ret = new Promise((resolve, reject) => {
-    let cleanup = () => {
-      renderer.removeListener('response', response);
-      renderer.removeListener('error', error);
-    };
-    let response = (event, resp) => {
-      cleanup();
-      resolve(resp);
-    };
-    let error = (event, err) => {
-      cleanup();
-      reject(err);
-    };
-
-    renderer.on('response', response);
-    renderer.on('error', error);
+    renderer.once('response', (_, res) => resolve(res));
+    renderer.once('error',    (_, err) => reject(err));
   });
 
   this.webContents.executeJavaScript(src);
@@ -243,26 +189,6 @@ Window.prototype.sendKey = function(ch){
 };
 
 /**
- * type {string, timeout}
- */
-Window.prototype.type = function(value, typeInterval){
-  let chars = String(value).split('');
-
-  let type = () => {
-    let ch = chars.shift();
-    if (ch === undefined) {
-      return Promise.resolve();
-    }
-
-    this.sendKey(ch);
-
-    return wait(typeInterval).then(type);
-  }
-
-  return type();
-};
-
-/**
  * screenshot
  */
 Window.prototype.screenshot = function(){
@@ -292,32 +218,6 @@ Window.prototype.continue = function(){
  * Utility Functions
  */
 
-// In most environments, loadURL handles this logic for us, but in some
-// it just hangs for unhandled protocols. Mitigate by checking ourselves.
-function canLoadProtocol(url) {
-  let protocol = urlFormat.parse(url).protocol || '';
-  protocol = protocol.replace(/:$/, '');
-
-  if (!protocol || KNOWN_PROTOCOLS.includes(protocol)) {
-    return Promise.resolve(true);
-  } else {
-    return new Promise((resolve, reject) => {
-      let done = (err, result) => {
-        if(err) reject(err);
-        if(!result){
-          reject({
-            message: 'navigation error',
-            code:    -1000,
-            details: 'unhandled protocol',
-            url:     url
-          });
-        }
-        resolve(result);
-      };
-      electron.protocol.isProtocolHandled(protocol, done);
-    });
-  }
-}
 // Format headers object for call to loadURL
 function toLoadURLOptions(headers){
   let httpReferrer = '';
@@ -335,10 +235,6 @@ function toLoadURLOptions(headers){
     loadUrlOptions.httpReferrer = httpReferrer;
   }
   return loadUrlOptions;
-}
-//Return a promise that resolves after {time}
-function wait(time){
-  return new Promise((resolve) => setTimeout(resolve, time));
 }
